Make test for firestream effect instead of all collision, fix collision

+++ projects/sample/GameStateEdit.cpp
@@ -74,7 +74,7 @@ GameStateEdit::GameStateEdit(SDL_Window* window): m_Window(window)
 		// TODO: Figure out how to handle resizes when resizing becomes a thing
 		int windowWidth, windowHeight;
 		SDL_GetWindowSize(m_Window, &windowWidth, &windowHeight);
-		std::unique_ptr<SystemCamera> camera = std::make_unique<SystemCamera>(windowWidth, windowHeight);
+		std::unique_ptr<SystemCamera> camera = std::make_unique<SystemCamera>(static_cast<float>(windowWidth), static_cast<float>(windowHeight));
 		m_SystemCamera = camera.get();
 		m_Systems.push_back(std::move(camera));
